+++ CHANGELOG.md
@@ -32,6 +32,9 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 - ADFineGrainedPasswordPolicy
   - New resource for creating and updating Fine Grained Password Policies for AD principal subjects.
     ([issue #584](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/584)).
+- ADDomainController
+  - Added support to demote domain controller when `ensure` is set to `Absent`.
+    ([issue #251](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/251))
 
 ### Changed